add initial manage package page
