+++ admin-portal/src/App.tsx
@@ -3,9 +3,9 @@ import { BrowserRouter, Route, Routes } from "react-router-dom";
 import { NavBar } from "./components/navbar/navbar";
 import { Home } from "./pages/home/home";
 import { NeedsUpload } from "./pages/needsupload/needsupload";
+import { AidPackage } from "pages/aidPackage/aidPackage";
+import { PackageDetails } from "./pages/packageDetails/packageDetails";
 import "./App.css";
-import NewAidPackage from "./pages/aidPackage/newAidPackage/NewAidPackage";
-import {PackageDetails} from "./pages/packageDetails/packageDetails";
 
 function App() {
   return (
@@ -17,7 +17,7 @@ function App() {
         <BrowserRouter>
           <Routes>
             <Route path="/" element={<Home />} />
-            <Route path="creation" element={<NewAidPackage />} />
+            <Route path="creation" element={<AidPackage />} />
             <Route path="needsupload" element={<NeedsUpload />} />
             <Route path="packages/:packageId" element={<PackageDetails />} />
           </Routes>
