+++ admin-portal/src/pages/aidPackage/aidPackage.css
@@ -1,7 +1,7 @@
 main {
   display: flex;
   flex-direction: column;
-  justify-content: start;
+  justify-content: flex-start;
   gap: 1rem;
   padding: 1rem;
 }
