+++ admin-portal/src/pages/aidPackage/manageAidPackages/manageAidPackages.tsx
@@ -1,0 +1,27 @@
+import { NeedsInfo } from "data/medical-needs.mock.data";
+import { NeedAssignments } from "../aidPackage";
+
+export function ManageAidPackages({
+  medicalNeeds,
+  needAssignments,
+  setNeedAssignments,
+}: {
+  medicalNeeds: Array<NeedsInfo>;
+  needAssignments: NeedAssignments;
+  setNeedAssignments: (needAssignments: NeedAssignments) => void;
+}) {
+  return (
+    <pre>
+      TODO:
+      <ul>
+        <li>create packages based on needAssignments passed in</li>
+        <li>allow editing package details</li>
+        <li>submit edited package details in a POST request</li>
+      </ul>
+      <h4>Needs Assignments</h4>
+      <code>{JSON.stringify(needAssignments, null, 2)}</code>
+      <h4>Needs List</h4>
+      <code>{JSON.stringify(medicalNeeds, null, 2)}</code>
+    </pre>
+  );
+}
